+++ source/poodinis/container.d
@@ -438,18 +438,6 @@ synchronized class DependencyContainer {
 	}
 
 	/**
-	 * Returns a global singleton instance of a dependency container.
-	 * Deprecated: create new instance with new keyword or implement your own singleton factory (method)
-	 */
-	deprecated public static shared(DependencyContainer) getInstance() {
-		static shared DependencyContainer instance;
-		if (instance is null) {
-			instance = new DependencyContainer();
-		}
-		return instance;
-	}
-
-	/**
 	 * Apply persistent registration options which will be used everytime register() is called.
 	 */
 	public void setPersistentRegistrationOptions(RegistrationOption options) {